Fix encryption mode handling and priority
If using a csv created from "List All in one table" on a radio reference system page, DE and De talkgroup modes are possible. This fix will force a DE (full-time encryption) mode talkgroup to be imported as E (Encrypted) mode and will force a De (part-time encryption) mode talkgroup to be imported as D (Digital) mode. Before these were both being imported as D (Digital) mode.

The priority row was off by 1 effectively causing all talkgroups to be imported with priority 3. This fix corrects that and also will default a talkgroup to priority 1 if no priority is defined, as is assumed by trunk-recorder.